TST: Reorganized tests from the Flight class into unit and integration.
